% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.mvCircular.R
\name{as.mvCircular}
\alias{as.mvCircular}
\title{Multivariate circular data conversor}
\usage{
as.mvCircular(x, control.circular = list(modulo = "2pi", zero = 0, template =
  "none", type = "angles", units = "radians", rotation = "counter"), ...)
}
\arguments{
\item{x}{The object to convert}

\item{control.ciruclar}{parameter for \link{circular::as.circular}}
}
\value{
Data frame with circular columns
}
\description{
Transforms a given object into a standard multivariate circular dataframe
}

