% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.circular.cov.R
\name{circular.cov}
\alias{circular.cov}
\title{Approximate circular covariance matrix}
\usage{
circular.cov(d, robust = F)
}
\arguments{
\item{d}{nxp data matrix}

\item{robust}{Use robust estimator for diagonal elements}
}
\value{
pxp circular covariance matrix
}
\description{
Computes the sample circular covariance matrix
}

