% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/empKL.circular.R
\name{empKL.circular}
\alias{empKL.circular}
\title{Empiric KL divergence}
\usage{
empKL.circular(a, b, n = nrow(a), m = nrow(b), k = 2)
}
\arguments{
\item{a}{Either a circular multivariate probability or a multivariate circular dataframe}

\item{b}{Either b circular multivariate probability or b multivariate circular dataframe}

\item{n}{If \code{a} is a probability, the number of samples from \code{a} to be used in the KL approximation}

\item{m}{If \code{b} is a probability, the number of samples from \code{b} to be used in the KL approximation}

\item{k}{The KL approximation is computed w.r.t. the distance to the \code{k}-th nearest neighbour of each sample}
}
\description{
Calculates a KL divergence approximation using distribution samples
and a simplified circular distance
}
\examples{
a <- rmvVonMises(10000,rep(0,3),rep(1,3),matrix(0,nrow=3,ncol=3))
b <- rmvVonMises(10000,rep(0,3),rep(1,3),matrix(0,nrow=3,ncol=3))
c <- rmvVonMises(10000,rep(2,3),rep(4,3),matrix(0,nrow=3,ncol=3))
div <- empKL.circular(a,b)
div$kl
div <- empKL.circular(a,c)
div$kl
}

