% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mvvonmises_naive.R
\name{fit_mvvonmises_naive}
\alias{fit_mvvonmises_naive}
\title{Fits a multivariate von Mises  distribution}
\usage{
fit_mvvonmises_naive(samples, mu = rep(0, ncol(samples)), kappa = rep(1,
  ncol(samples)), lambda = matrix(0, ncol = ncol(samples), nrow =
  ncol(samples)), verbose = -1, prec = 1e+05, tolerance = 1e-07,
  mprec = 30, retry = 10)
}
\arguments{
\item{samples}{Matrix or df of circular samples}

\item{mu}{Initial value for mean vector. It should be Circular.}

\item{kappa}{Initial value for concentration vector}

\item{lambda}{Initial value for lambda matrix}

\item{verbose}{Verbose degree parameter. -1 for silent mode.}

\item{prec}{Precision parameter of the LBFGS-B method}

\item{tolerance}{Tolerance parameter of the LBFGS-B mehtod}

\item{mprec}{Number of steps that the pseudoinverse calculation remembers}

\item{retry}{Number of retries (Convergence can fail sometimes)}

\item{type}{Wheter to use optimized expresion or regular}
}
\value{
List with mu, kappa and lambda parameters, data samples and loss function value. If method don't reach convergence loss will be NA
}
\description{
Given a set of multivariate circular samples, compute ml mu, kappa and lambda parameters
based on the pseudolikelihood function. Descent is made using lbfgs-b algorithm.
}

