% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geomedian.circular.R
\name{geomedian.circular}
\alias{geomedian.circular}
\title{Cicular geometric median}
\usage{
geomedian.circular(data, tol = 0.001, maxiter = 1e+05)
}
\arguments{
\item{data}{Circular dataframe or matrix}

\item{tol}{geometric median tolerance}

\item{maxiter}{maximum number of iterations}
}
\description{
Compute geometric circular median based on the simplified circular distance
}
\examples{
geomedian.circular( rmvVonMises(1000,rep(0,3),rep(1,3),matrix(0,nrow=3,ncol=3)) )
}

