% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.mvCircular.R
\name{is.mvCircular}
\alias{is.mvCircular}
\title{Multivariate circular data test}
\usage{
is.mvCircular(x)
}
\arguments{
\item{x}{The object to test}
}
\value{
True/False
}
\description{
Checks if an object is multivariate circular data
}

