\name{mvCircular-package}
\alias{mvCircular-package}
\alias{mvCircular}
\docType{package}
\title{
\packageTitle{mvCircular}
}
\description{
\packageDescription{mvCircular}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mvCircular}
\packageIndices{mvCircular}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mvCircular}

Maintainer: \packageMaintainer{mvCircular}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
