% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genProbDist.R
\name{mvCircularProbDist}
\alias{Fval}
\alias{circCor}
\alias{circMarginal}
\alias{circMarginalConcentration}
\alias{circMarginalMean}
\alias{contourPlot}
\alias{fval}
\alias{getSamples}
\alias{kl}
\alias{mvCircularProbDist}
\alias{normalize}
\alias{plot.mvCircularProbDist}
\alias{torusPlot}
\alias{torusPlot.mvCircularProbDist}
\title{Multivariate circular distribution}
\usage{
mvCircularProbDist(dim, ...)

getSamples(obj, n, ...)

fval(obj, x, ...)

Fval(obj, x, ...)

normalize(obj, ...)

circMarginal(obj, x, i, ...)

circCor(obj, i, j, ...)

circMarginalMean(obj, i, ...)

circMarginalConcentration(obj, i, ...)

contourPlot(obj, ...)

torusPlot(obj, ...)

\method{torusPlot}{mvCircularProbDist}(obj, npoints = 10000, ...)

\method{plot}{mvCircularProbDist}(obj, data = NULL, n = 100, ...,
  color = "red", cex.params = 2, rose.bins = 25, cex.sigma = 2,
  pch = 18)

kl(obj, q, ...)
}
\arguments{
\item{dim}{Number of variables}

\item{obj}{A multivariate circular distribution object}

\item{n}{The number of samples to return}

\item{x}{A circular point (p-vector)}

\item{i}{Component index}

\item{j}{Component index}

\item{npoints}{Number of points to evaluate}

\item{data}{Data to plot in the lower triagonal segment. Usually data used for fitting.}

\item{color}{Plot main color}

\item{cex.params}{Parameters font size}

\item{rose.bins}{Number of bins in the rose plot}

\item{cex.sigma}{Upper triangular dep. font size}

\item{pch}{Point character to use in the 2D plots}

\item{\dots}{(\code{mvCircularProbDist}) Named list with aditional parameters to include in the object}

\item{\dots}{(\code{torusPlot}) Additional params for \code{feval}}

\item{\dots}{(plot) Additional arguments for \link{circMarginal}}
}
\value{
\code{getSamples} returns a circular dataframe with n samples from the given circular distribution

\code{fval} evaluates the density function in the point \code{x}

\code{Fval} returns the cumulative distribution function in the point \code{x}

\code{normalize} returns the object with the normalization term set

\code{circMarginal} returns the \code{i}-th compontent marginal density function evaluated at \code{x}

\code{circCor} returns the dependency between \code{i}-th and \code{j}-th compontents

\code{circMarginalMean} returns the marginal mean of the \code{i}-th compontent

\code{circMarginalConcentration} returns the marginal concentration of the \code{i}-th compontent

\code{kl} returns the KL divergence value between obj and q
}
\description{
Generic defnition for a multivariate ciruclar distribution

\code{contourPlot} creates a 2D plot where the level based on on the density function value. 
Only available for bivariate distributions

\code{torusPlot} creates a 3D interactive plot using \link{threejs} library. The surface of a torus is coloured
based on the density function value. Only available for bivariate distributions
}
\examples{
 obj <- mvVonMises(c(3*pi/2, pi/2), rep(1.5,2), matrix( c(0,2,2,0) ,ncol=2,nrow=2) )
 obj <- normalize(obj)
 torusPlot(obj)
 
 @export
 obj <- mvVonMises(c(3*pi/2, pi/2, 0, 0), c(1.5,2,1,5), matrix( c(0,2,-1,0, 2,0,1,0, -1,1,0,1, 0,0,1,0) ,ncol=4,nrow=4) )
 plot(obj)

}
\author{
Luis Rodriguez Lujan
}
\keyword{circular}
\keyword{multivariate}

