% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvNormal.R
\name{mvNormal}
\alias{circMarginal.mvNormal}
\alias{dmvNormal}
\alias{fval.mvNormal}
\alias{getSamples.mvNormal}
\alias{mvNormal}
\alias{mvNormal.fit}
\alias{rmvNormal}
\title{Multivariate normal distribution}
\usage{
mvNormal(mu, sigma, ...)

mvNormal.fit(samples, zero.threshold = 0.01, ...)

dmvNormal(x, mu, sigma, ...)

rmvNormal(n, mu, sigma, ...)

\method{getSamples}{mvNormal}(obj, n, ...)

\method{fval}{mvNormal}(obj, x, k = 10, ...)

\method{circMarginal}{mvNormal}(obj, x, i)
}
\arguments{
\item{mu}{Circular mean vector}

\item{sigma}{a positive-definite symmetric matrix that specifies covariance matrix}

\item{...}{(\code{dmvNormal}) extra arguments for \code{\link{mvtnorm::dmvnorm}}}

\item{samples}{Matrix or DF with multivariate samples}

\item{zero.threshold}{Any sigma value that verifies that \code{abs(x) < zero.threshold } is returned as zero}

\item{x}{The point to evaluate}

\item{n}{Number of samples to generate}

\item{\dots}{(Constructor) Named list with additional attributes to add to the object}

\item{...}{(\code{rmvNormal}) Additional parameters for \code{\link{mvtnorm::rmvnorm} }}
}
\value{
\code{mvNormal} returns a mvNormal object

\code{dmvNormal} returns the density function evaluated at \code{x}

\code{rmvNormal} returns a multivariate circular dataframe with \code{n} 
samples from a normal distribution
}
\description{
Fit method uses data matrix or dataframe \code{samples} to compute the ML parameters of the distribution

\code{dmvNormal} computes multivariate normal densitiy function approximately. The precission is controled by
\code{k}, the number of points to evaluate per dimension. The total number of points will be  (k+1) ^ ndim (Z-lattice)
}
\examples{
mvNormal(rep(0,3), diag(3) )


}
\author{
Luis Rodriguez Lujan
}
\seealso{
\code{\link{mvCircularProbDist}}
}
\keyword{multivariate}
\keyword{normal}

