% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvvm_loglik.R
\name{mvVonMises.logplik.pen}
\alias{mvVonMises.logplik.pen}
\title{#@param par parameter vector (length p + p(p-1)/2 ) in kappa - upper.tri(lambda) order}
\usage{
mvVonMises.logplik.pen(par, phi = NULL, H = NULL, alpha = NULL, data.s,
  data.c)
}
\description{
#@param par parameter vector (length p + p(p-1)/2 ) in kappa - upper.tri(lambda) order
}

