% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmvvonmises_rs.r
\name{rmvvonmises_rs}
\alias{rmvvonmises_rs}
\title{Multivariate von Mises sampler}
\usage{
rmvvonmises_rs(n, mu, kappa, lambda, type = c("auto", "gibbs", "rejection"),
  burnin = 0.5, burnin.tol = 0.1, burnin.chains = chains,
  thinning = 0.01, chains = 5, alpha = rep(1/length(mu), length(mu)),
  alphaPrec = 4, seed = runif(64, 0, .Machine$integer.max), finalSize = T,
  verbose = F)
}
\arguments{
\item{n}{Number of samples to be generated}

\item{mu}{Mean circular vector}

\item{kappa}{Concentration vector}

\item{lambda}{Dependency matrix}

\item{type}{\code{auto} for automatic sampler selection, \code{gibbs} for gibbs sampler and \code{rejection} for rejection sampling}

\item{burnin}{(Only for Gibbs) \code{0} to deactivate burning. Between \code{0} and \code{1} a ratio of burn-in iterations w.r.t. n.
Otherwise the number of burn-in iterations. If \code{burnin} is a negative number, adaptative burn-in will be performed}

\item{burnin.tol}{(Only for Gibbs) Convergece threshold for iterative burn-in based on gelman mpsrf \code{\link{coda::gelman.diag}}}

\item{burnin.chains}{(Only for Gibbs) Number of chains to be used in the iterative burn-in}

\item{thinning}{(Only for Gibbs) Distance between accepted samples in the Gibbs samples.}

\item{chains}{(Only for Gibbs) Number of chains to merge for the final result.}

\item{alpha}{(Only for Gibbs) Component update probability in each Gibbs iteration (should sum 1)}

\item{alphaPrec}{(Only for Gibbs) Alpha lookup table size \code{10^alphaPrec}}

\item{seed}{Seed to be used in the random number generator}

\item{finalSize}{True ensures that the final output will contain \code{n} samples}

\item{verbose}{True for verbose output}
}
\description{
Multivariate von Mises sampler
}
\examples{
rmvvonmises_rs(100,rep(0,3),rep(1,3),matrix(rep(0,9),nrow=3,ncol=3))
}

